New animated intermission added
The VERY GOOD !! intermission has been imported. You will see it at the very beginning. Up to you decide where to use it and what to show. 
All fonts are loaded in VRAM. Numbers start at 216 with "0" and ends at 215 with "9", letters start at 226 with "A" and ends at 251 with "Z". The 4 tiles at 252-255 are ",-.!"