Glitch solution (maybe..)
In this version I've removed the code that was filling the copy of the SAT in RAM from all the intermissions and the extra mr do screens. 
The screen transitions are a bit dirty now, but I can fix them later in another way. 
Apparently the glitch concerning the "apple" sprite disappearing and the chomper in the scorebar does not appear anymore (I did only a few of play, more testing is required).

Never the less I'm very sceptical about this solution, because, when a level change occurs, all apples should be restored in the SAT in RAM during the level set up, reverting any "error" in there.
So something should be occurred...